+++ src/MonaCtrl/Camera.h
@@ -35,14 +35,14 @@ public:
 
 
    // Set viewing frustum.
-   void setFrustum(GLfloat angle, GLfloat aspect,
-                   GLfloat znear, GLfloat zfar);
+   void setFrustum(GLdouble angle, GLdouble aspect,
+                   GLdouble znear, GLdouble zfar);
 
    // Set camera view.
-   void setView(GLfloat *eye, GLfloat *lookat, GLfloat *up);
+   void setView(GLdouble *eye, GLdouble *lookat, GLdouble *up);
 
    // Update camera view.
-   void updateView(GLfloat *eye, GLfloat *lookat, GLfloat *up);
+   void updateView(GLdouble *eye, GLdouble *lookat, GLdouble *up);
 
    // Place camera.
    void place();
@@ -52,23 +52,23 @@ public:
    void setCameraSpringiness(int size);
 
    // Is local point inside frustum (visible)?
-   bool pointInFrustum(GLfloat *localPoint);
+   bool pointInFrustum(GLdouble *localPoint);
 
 private:
 
    // Camera vector history for springiness calculation.
    struct CameraHistory
    {
-      GLfloat e[3];                               // eye
-      GLfloat f[3];                               // forward
-      GLfloat u[3];                               // up
+      GLdouble e[3];                               // eye
+      GLdouble f[3];                               // forward
+      GLdouble u[3];                               // up
    }
        *cameraHistory;
    int cameraHistorySize;
    int cameraHistoryIndex;
 
    // Frustum bounding planes.
-   GLfloat frustum[6][4];
+   GLdouble frustum[6][4];
 
    // Extract frustum.
    void extractFrustum();
